// backends/payload/src/payload-types.ts
export interface User {
  id: string;
  displayName: string;
  groupMutationVersion?: number | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  resetPasswordRequestedAt?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "groups".
 */
export interface Group {
  id: string;
  owner: string | User;
  name: string;
  rank: string;
  sharingEnabled?: boolean | null;
  shareVersion?: number | null;
  revision: number;
  updatedAt: string;
  createdAt: string;
}